+++ src/main/java/com/sefaz/sistemaficha/services/FichaResumidaViewService.java
@@ -21,12 +21,12 @@ public class FichaResumidaViewService {
 		return list;
 	}
 
-	public FichaResumidaView findById(Integer id) {
+	public FichaResumidaView findById(Long id) {
 		Optional<FichaResumidaView> ficha = repository.findById(id);
 		return ficha.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
-	public List<FichaResumidaView> findAllPorAnotacaoId(Integer id){
+	public List<FichaResumidaView> findAllPorAnotacaoId(Long id){
 		List<FichaResumidaView> list = repository.findAllPorAnotacaoId(id);
 		return list;
 	}
