+++ src/main/java/com/sefaz/sistemaficha/services/exceptions/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package com.sefaz.sistemaficha.services.exceptions;
 public class ResourceNotFoundException extends RuntimeException{
 	private static final long serialVersionUID = 1L;
 
-	public ResourceNotFoundException (Integer id) {
+	public ResourceNotFoundException (Long id) {
 		super("Resource not found. Id: " + id);
 	}
 }
